requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

arrow==1.2.3; python_version >= '3.6'

asgiref==3.6.0; python_version >= '3.7'

binaryornot==0.4.4

certifi==2023.5.7; python_version >= '3.6'

chardet==5.1.0; python_version >= '3.7'

charset-normalizer==3.1.0; python_version >= '3.7'

click==8.1.3; python_version >= '3.7'

colorama==0.4.6; platform_system == 'Windows'

cookiecutter==2.1.1

django-browser-reload==1.8.0; python_version >= '3.7'

django-jazzmin==2.6.0

django-shortuuidfield==0.1.3

django-tailwind==3.5.0

django==4.2.1

djangorestframework==3.14.0

idna==3.4; python_version >= '3.5'

jinja2-time==0.2.0

jinja2==3.1.2; python_version >= '3.7'

markupsafe==2.1.2; python_version >= '3.7'

pillow==9.5.0

python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

python-slugify==8.0.1; python_version >= '3.7'

pytz==2023.3

pyyaml==6.0; python_version >= '3.6'

requests==2.30.0; python_version >= '3.7'

shortuuid==1.0.11; python_version >= '3.5'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sqlparse==0.4.4; python_version >= '3.5'

text-unidecode==1.3

tzdata==2023.3; sys_platform == 'win32'

urllib3==2.0.2; python_version >= '3.7'

